Moved config parameters to separate class. Merged un/authenticated HTTP requests into a single class.
