+++ src/test/java/bio/terra/endpoints/CLITestingConfig.java
@@ -1,0 +1,12 @@
+package bio.terra.endpoints;
+
+public class CLITestingConfig {
+    //public static final String dataRepoURL = "http://localhost:8080/";
+    //public static final String clientSecretsFilePath = null;
+    //public static final String dataRepoURL = "https://jade.datarepo-dev.broadinstitute.org/";
+    //public static final String clientSecretsFilePath = "/tmp/jadecli_client_secret.json";
+    public static final String dataRepoURL = "https://jade-dd.datarepo-dev.broadinstitute.org/";
+    public static final String clientSecretsFilePath = "/tmp/jadecli_client_secret.json";
+
+    public static String dirName = "src/test/resources/CLICommandTests/";
+}
